+++ packages/control-property-editor/test/unit/slice.test.ts
@@ -402,7 +402,8 @@ describe('main redux slice', () => {
                     reloadApplication()
                 )
             ).toStrictEqual({
-                fileChanges: []
+                fileChanges: [],
+                isAppLoading: true
             });
         });
     });