+++ packages/preview-middleware/test/unit/base/flp.test.ts
@@ -391,6 +391,98 @@ describe('FlpSandbox', () => {
         test('no route for custom init', async () => {
             await server.get('/test/integration/opaTests.qunit.js').expect(404);
         });
+
+        test('default without testsuite', async () => {
+            await server.get('/test/testsuite.qunit.html').expect(404);
+            await server.get('/test/testsuite.qunit.js').expect(404);
+        });
+    });
+
+    describe('router with test suite', () => {
+        let server!: SuperTest<Test>;
+
+        beforeAll(async () => {
+            const flp = new FlpSandbox(
+                {
+                    flp: {
+                        apps: [
+                            {
+                                target: '/yet/another/app'
+                            }
+                        ]
+                    },
+                    test: [
+                        {
+                            framework: 'QUnit'
+                        },
+                        {
+                            framework: 'OPA5'
+                        },
+                        {
+                            framework: 'Testsuite'
+                        }
+                    ]
+                },
+                mockProject,
+                mockUtils,
+                logger
+            );
+            const manifest = JSON.parse(readFileSync(join(fixtures, 'simple-app/webapp/manifest.json'), 'utf-8'));
+            await flp.init(manifest);
+
+            const app = express();
+            app.use(flp.router);
+
+            server = await supertest(app);
+        });
+
+        test('default with testsuite', async () => {
+            const response = await server.get('/test/testsuite.qunit.html').expect(200);
+            expect(response.text).toMatchSnapshot();
+        });
+
+        test('default with testsuite and init testsuite.qunit.js', async () => {
+            const response = await server.get('/test/testsuite.qunit.js').expect(200);
+            expect(response.text).toMatchSnapshot();
+        });
+    });
+
+    describe('router with test suite (negative)', () => {
+        let server!: SuperTest<Test>;
+
+        beforeAll(async () => {
+            const flp = new FlpSandbox(
+                {
+                    flp: {
+                        apps: [
+                            {
+                                target: '/yet/another/app'
+                            }
+                        ]
+                    },
+                    test: [
+                        {
+                            framework: 'Testsuite'
+                        }
+                    ]
+                },
+                mockProject,
+                mockUtils,
+                logger
+            );
+            const manifest = JSON.parse(readFileSync(join(fixtures, 'simple-app/webapp/manifest.json'), 'utf-8'));
+            await flp.init(manifest);
+
+            const app = express();
+            app.use(flp.router);
+
+            server = await supertest(app);
+        });
+
+        test('no testsuite w/o test frameworks', async () => {
+            await server.get('/test/testsuite.qunit.html').expect(404);
+            await server.get('/test/testsuite.qunit.js').expect(404);
+        });
     });
 
     describe('router - existing FlpSandbox', () => {