+++ packages/system-access/test/unit/base/connect.test.ts
@@ -4,9 +4,10 @@ import { mockedStoreService, mockIsAppStudio, mockListDestinations, mockReadFile
 import type { Destination } from '@sap-ux/btp-utils';
 import type { AbapTarget } from '../../../src/types';
 import prompts from 'prompts';
+import { AuthenticationType } from '@sap-ux/store';
 import nock from 'nock';
 
-describe('service', () => {
+describe('connect', () => {
     const logger = new ToolsLogger({ transports: [new NullTransport()] });
     const target = {
         url: 'http://target.example',
@@ -37,11 +38,17 @@ describe('service', () => {
             });
 
             test('prompt credentials if not available in store', async () => {
-                prompts.inject([username, password]);
+                prompts.inject([AuthenticationType.Basic, username, password]);
                 const provider = await createAbapServiceProvider(target, undefined, true, logger);
                 expect(provider).toBeDefined();
                 expect(process.env.FIORI_TOOLS_USER).toBe(username);
                 expect(process.env.FIORI_TOOLS_PASSWORD).toBe(password);
+            });
+
+            test('prompt S/4Cloud system if not available in store', async () => {
+                prompts.inject([AuthenticationType.ReentranceTicket]);
+                const provider = await createAbapServiceProvider({ ...target }, undefined, true, logger);
+                expect(provider).toBeDefined();
             });
 
             test('invalid target', async () => {
@@ -105,7 +112,7 @@ describe('service', () => {
             test('throw error when cloud system read from store but cloud target is not specified in params', async () => {
                 mockedStoreService.read.mockResolvedValueOnce(credentials);
                 try {
-                    await createAbapServiceProvider({ ...target, scp: false }, undefined, true, logger);
+                    await createAbapServiceProvider({ ...target, scp: false }, undefined, false, logger);
                     fail('Should have thrown an error');
                 } catch (error) {
                     expect(error.message).toBe('This is an ABAP Cloud system, please correct your configuration.');