+++ packages/telemetry/test/tools-suite-telemetry/index.test.ts
@@ -487,10 +487,7 @@ describe('Tools Suite Telemetry Tests', () => {
     it('common telemetry property: adaptation project', async () => {
         memfs.vol.fromNestedJSON(
             {
-                ['/project1/.adp/config.json']: fs.readFileSync(
-                    join(__dirname, 'fixtures/valid-adaptation/.adp/config.json'),
-                    'utf-8'
-                ),
+                ['/project1/ui5.yaml']: fs.readFileSync(join(__dirname, 'fixtures/valid-adaptation/ui5.yaml'), 'utf-8'),
                 ['/project1/webapp/manifest.appdescr_variant']: fs.readFileSync(
                     join(__dirname, 'fixtures/valid-adaptation/webapp/manifest.appdescr_variant'),
                     'utf-8'
@@ -512,7 +509,48 @@ describe('Tools Suite Telemetry Tests', () => {
             appstudio: false,
             'cmn.appstudio': false,
             'cmn.template': '',
-            'cmn.toolsId': 'NO_TOOLS_ID',
+            'cmn.deployTarget': 'NO_DEPLOY_CONFIG',
+            'cmn.odataSource': 'ABAP', // New project-access implementation assume ABAP data source if its not CAP
+            'cmn.devspace': '',
+            'cmn.internalFeatures': 'external',
+            internalVsExternal: 'external',
+            'cmn.nodeVersion': expect.any(String),
+            'cmn.templateId': '@sap/generator-adaptation-project',
+            'cmn.templateVersion': '1.1.57',
+            'cmn.toolsId': '421a7e6d-4507-4f66-9369-f80bdd3c6877',
+            'cmn.appLanguage': '',
+            'cmn.applicationType': 'Fiori Adaptation'
+        });
+    });
+
+    it('common telemetry property: invalid adaptation project', async () => {
+        memfs.vol.fromNestedJSON(
+            {
+                ['/project1/ui5.yaml']: fs.readFileSync(
+                    join(__dirname, 'fixtures/invalid-adaptation/ui5.yaml'),
+                    'utf-8'
+                ),
+                ['/project1/webapp/manifest.appdescr_variant']: fs.readFileSync(
+                    join(__dirname, 'fixtures/invalid-adaptation/webapp/manifest.appdescr_variant'),
+                    'utf-8'
+                ),
+                ['/project1/package.json']: fs.readFileSync(
+                    join(__dirname, 'fixtures/invalid-adaptation/package.json'),
+                    'utf-8'
+                )
+            },
+            '/'
+        );
+
+        isAppStudioMock.mockReturnValue(false);
+        const commonProperties = await processToolsSuiteTelemetry({
+            appPath: '/project1'
+        });
+
+        expect(commonProperties).toEqual({
+            appstudio: false,
+            'cmn.appstudio': false,
+            'cmn.template': '',
             'cmn.deployTarget': 'NO_DEPLOY_CONFIG',
             'cmn.odataSource': 'ABAP', // New project-access implementation assume ABAP data source if its not CAP
             'cmn.devspace': '',
@@ -521,6 +559,7 @@ describe('Tools Suite Telemetry Tests', () => {
             'cmn.nodeVersion': expect.any(String),
             'cmn.templateId': '',
             'cmn.templateVersion': '',
+            'cmn.toolsId': 'NO_TOOLS_ID',
             'cmn.appLanguage': '',
             'cmn.applicationType': 'Fiori Adaptation'
         });