+++ examples/simple-generator/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @sap-ux/generator-simple-fe
 
+## 0.3.62
+
+### Patch Changes
+
+-   Updated dependencies [d6b68b0e]
+    -   @sap-ux/fiori-elements-writer@0.19.8
+
+## 0.3.61
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@0.19.7
+-   @sap-ux/fiori-freestyle-writer@0.22.7
+-   @sap-ux/axios-extension@1.12.4
+-   @sap-ux/system-access@0.3.30
+
+## 0.3.60
+
+### Patch Changes
+
+-   Updated dependencies [96de8f25]
+    -   @sap-ux/system-access@0.3.30
+
 ## 0.3.59
 
 ### Patch Changes