+++ packages/fiori-generator-shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/fiori-generator-shared
 
+## 0.0.7
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.20.1
+
 ## 0.0.6
 
 ### Patch Changes