+++ packages/odata-service-writer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/odata-service-writer
 
+## 0.19.2
+
+### Patch Changes
+
+-   Updated dependencies [3684195d]
+    -   @sap-ux/ui5-config@0.22.5
+    -   @sap-ux/mockserver-config-writer@0.5.3
+
 ## 0.19.1
 
 ### Patch Changes