+++ packages/system-access/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/system-access
 
+## 0.3.30
+
+### Patch Changes
+
+-   96de8f25: FIX: prompt user for authentication type if not known when generating an adaptation project
+
 ## 0.3.29
 
 ### Patch Changes