+++ packages/ui5-library-reference-inquirer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/ui5-library-reference-inquirer
 
+## 0.2.24
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.20.1
+
 ## 0.2.23
 
 ### Patch Changes