+++ packages/ui5-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/ui5-proxy-middleware
 
+## 1.3.10
+
+### Patch Changes
+
+-   Updated dependencies [3684195d]
+    -   @sap-ux/ui5-config@0.22.5
+
 ## 1.3.9
 
 ### Patch Changes