+++ packages/yaml/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/yaml
 
+## 0.14.2
+
+### Patch Changes
+
+-   3684195d: adds authenticationType to fiori tools proxy backend and ability to add inline comments to yaml nodes
+
 ## 0.14.1
 
 ### Patch Changes